Refactor LLVMValue and LLVMScalar header into Model folder
