+++ include/caffeine/Solver/Solver.h
@@ -9,7 +9,7 @@
 #include "caffeine/ADT/Ref.h"
 #include "caffeine/IR/Value.h"
 #include "caffeine/Interpreter/AssertionList.h"
-#include "caffeine/Interpreter/Value.h"
+#include "caffeine/Model/Value.h"
 
 namespace caffeine {
